requirements.txt
crewai
groq
streamlit
reportlab
python-dotenv
langchain-groq
langchain
langchain-community